Split workbench entity tests into per-config files to match docs convention (#255)
docs/frontend.md specifies tests mirror source structure one-to-one, and the
repo's dominant pattern mirrors source subfolders into test subfolders. Split
the consolidated entities.test.ts into per-entity files under tests/.../entities/
(zone, enemy, challenge, item-mod) accordingly, with no change in coverage.
